* renamed packages,artifacts,groups,permissions, tool-id,css-classes,.. * updated wicket version * removed some debug logging * fixed minor things which caught my eye while I was doing all the other stuff
