--- api/src/java/net/stellers/sakai/academicsessionadmin/exceptions/DuplicateKeyException.java
@@ -1,11 +1,0 @@
-package net.stellers.sakai.academicsessionadmin.exceptions;
-
-public class DuplicateKeyException extends Exception {
-
-	private static final long serialVersionUID = 1L;
-	public DuplicateKeyException (String msg){
-		super(msg);
-	}
-	
-	
-}
