--- api/src/java/net/stellers/sakai/academicsessionadmin/exceptions/NoSuchKeyException.java
@@ -1,12 +1,0 @@
-package net.stellers.sakai.academicsessionadmin.exceptions;
-
-public class NoSuchKeyException extends Exception {
-
-	private static final long serialVersionUID = 1L;
-	
-	public NoSuchKeyException(String msg){
-		super(msg);
-	}
-
-	
-}
