+++ api/src/java/org/sakaiproject/acadtermmanage/exceptions/DuplicateKeyException.java
@@ -1,0 +1,11 @@
+package org.sakaiproject.acadtermmanage.exceptions;
+
+public class DuplicateKeyException extends Exception {
+
+	private static final long serialVersionUID = 1L;
+	public DuplicateKeyException (String msg){
+		super(msg);
+	}
+	
+	
+}
