+++ api/src/java/org/sakaiproject/acadtermmanage/exceptions/NoSuchKeyException.java
@@ -1,0 +1,12 @@
+package org.sakaiproject.acadtermmanage.exceptions;
+
+public class NoSuchKeyException extends Exception {
+
+	private static final long serialVersionUID = 1L;
+	
+	public NoSuchKeyException(String msg){
+		super(msg);
+	}
+
+	
+}
