+++ impl/pom.xml
@@ -2,22 +2,22 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 	<parent>
-    <artifactId>sakai-term-admin</artifactId>
-    <groupId>net.stellers.sakai</groupId>
-    <version>1.11.x-SNAPSHOT</version>
-  </parent>
+	    <groupId>org.sakaiproject.acadtermmanage</groupId>	
+    	<artifactId>sakai-acadtermmanage</artifactId>
+    	<version>11.4-SNAPSHOT</version>
+    	<relativePath>../pom.xml</relativePath>    	
+  	</parent>
 	
-	<name>sakai-term-admin - Implementation</name>
-	<artifactId>sakai-term-admin-impl</artifactId>
-
+	<name>sakai-acadtermmanage - Implementation</name>
+	<artifactId>sakai-acadtermmanage-impl</artifactId>
 	<packaging>jar</packaging>
 	
   	<dependencies>
 
     	<!-- project dependencies -->
     	<dependency>
-			<groupId>net.stellers.sakai</groupId>		
-			<artifactId>sakai-term-admin-api</artifactId>			
+			<groupId>org.sakaiproject.acadtermmanage</groupId>		
+			<artifactId>sakai-acadtermmanage-api</artifactId>			
 		</dependency>
