+++ pack/src/webapp/WEB-INF/components.xml
@@ -3,24 +3,24 @@
 
 <beans>	
 	<!-- Project logic   -->
-	<bean id="net.stellers.sakai.academicsessionadmin.logic.AcademicSessionLogic"
-    	class="net.stellers.sakai.academicsessionadmin.logic.impl.AcademicSessionLogicImpl"
+	<bean id="org.sakaiproject.acadtermmanage.logic.AcademicSessionLogic"
+    	class="org.sakaiproject.acadtermmanage.logic.impl.AcademicSessionLogicImpl"
     	init-method="init">    	
-    	<property name="dao" ref="net.stellers.sakai.academicsessionadmin.dao.AcademicSessionDao" />
-  		<property name="as_sakaiProxy" ref="net.stellers.sakai.academicsessionadmin.logic.AcademicSessionSakaiProxy" />    		
-    	<property name="cache" ref="net.stellers.sakai.academicsessionadmin.sakai.academicsession_cache" />	    	
+    	<property name="dao" ref="org.sakaiproject.acadtermmanage.dao.AcademicSessionDao" />
+  		<property name="as_sakaiProxy" ref="org.sakaiproject.acadtermmanage.logic.AcademicSessionSakaiProxy" />    		
+    	<property name="cache" ref="org.sakaiproject.acadtermmanage.sakai.academicsession_cache" />	    	
     </bean>
         
     <!--  Project DAOs -->
-    <bean id="net.stellers.sakai.academicsessionadmin.dao.AcademicSessionDao"
-		class="net.stellers.sakai.academicsessionadmin.dao.impl.AcademicSessionDaoImpl"
+    <bean id="org.sakaiproject.acadtermmanage.dao.AcademicSessionDao"
+		class="org.sakaiproject.acadtermmanage.dao.impl.AcademicSessionDaoImpl"
 		init-method="init">		
-		<property name="cm_api" ref="net.stellers.sakai.academicsessionadmin.logic.AcademicSessionSakaiProxy" />
+		<property name="cm_api" ref="org.sakaiproject.acadtermmanage.logic.AcademicSessionSakaiProxy" />
 	</bean>	    
 		
 	<!--  SakaiProxy -->
- 	<bean id="net.stellers.sakai.academicsessionadmin.logic.AcademicSessionSakaiProxy"
-    	class="net.stellers.sakai.academicsessionadmin.logic.impl.AcademicSessionSakaiProxyImpl"
+ 	<bean id="org.sakaiproject.acadtermmanage.logic.AcademicSessionSakaiProxy"
+    	class="org.sakaiproject.acadtermmanage.logic.impl.AcademicSessionSakaiProxyImpl"
         init-method="init">        
         <property name="cmService" ref="org.sakaiproject.coursemanagement.api.CourseManagementService" />
         <property name="cmAdmin" ref="org.sakaiproject.coursemanagement.api.CourseManagementAdministration" />  
@@ -35,13 +35,13 @@
 	</bean>
 	
 	<!-- Cache: ttl=1 hour -->
-    <bean id="net.stellers.sakai.academicsessionadmin.sakai.academicsession_cache"
+    <bean id="org.sakaiproject.acadtermmanage.sakai.academicsession_cache"
         class="org.springframework.cache.ehcache.EhCacheFactoryBean">
         <property name="cacheManager">
             <ref bean="org.sakaiproject.memory.api.MemoryService.cacheManager"/>
         </property>
         <property name="cacheName">
-            <value>net.stellers.sakai.academicsessionadmin.sakai.academicsession_cache</value>
+            <value>org.sakaiproject.acadtermmanage.sakai.academicsession_cache</value>
         </property>
         <property name="maxElementsInMemory" value="200" />
         <property name="diskPersistent" value="false" />
