--- tool/src/java/net/stellers/sakai/academicsessionadmin/tool/Constants.java
@@ -1,16 +1,0 @@
-package net.stellers.sakai.academicsessionadmin.tool;
-
-public interface Constants {
-	public static final String PROP_EID="eid";
-	public static final String PROP_START="startDate";
-	public static final String PROP_END="endDate";
-	public static final String PROP_DESC="description";
-	public static final String PROP_TITLE="title";
-	public static final String PROP_CURRENT="current";
-	
-	public static final String LABEL_BUTTON_ADD ="lbl_button_add";
-	public static final String LABEL_BUTTON_SAVE ="lbl_button_save";
-	public static final String LABEL_BUTTON_CANCEL ="lbl_button_cancel";
-	public static final String LABEL_EDIT ="lbl_edit";
-	
-}
