--- tool/src/java/net/stellers/sakai/academicsessionadmin/tool/pages/BasePage.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
-<html xmlns:wicket="http://wicket.apache.org/dtds.data/wicket-xhtml1.4-strict.dtd" >
- 
-<head>
-	<title><wicket:message key="page.title" /></title>
-</head>     
-<body>
-	<div class="portletBody">	
-		<div class="fuapi-feedbackpanel" wicket:id="feedback">feedbackmessages will be put here</div>					
-		<wicket:child />	
-	</div>
-</body>
-</html>
