--- tool/src/java/net/stellers/sakai/academicsessionadmin/tool/wicketstuff/ActionPanel.html
@@ -1,3 +1,0 @@
-<wicket:panel xmlns:wicket="http://wicket.apache.org">
-<a href="#" class="fu_asa_action_link" wicket:id="id_action_link">edit</a>
-</wicket:panel>
