+++ tool/src/java/org/sakaiproject/acadtermmanage/tool/AcademicSessionAdminApplication.properties
