+++ tool/src/java/org/sakaiproject/acadtermmanage/tool/wicketstuff/ActionPanel.html
@@ -1,0 +1,3 @@
+<wicket:panel xmlns:wicket="http://wicket.apache.org">
+<a href="#" class="sakai.acadtermmanage-action_link" wicket:id="id_action_link">edit</a>
+</wicket:panel>
