+++ tool/src/webapp/WEB-INF/web.xml
@@ -6,21 +6,21 @@
 	 id="academicSessionApp"
 	 >
 
-	<display-name>sakai.term-admin</display-name>
+	<display-name>sakai.acadtermmanage</display-name>
 
 	<!--  This application's servlet class -->
 	<servlet>
-		<servlet-name>sakai.term-admin</servlet-name>
+		<servlet-name>sakai.acadtermmanage</servlet-name>
 	    <servlet-class>org.apache.wicket.protocol.http.WicketServlet</servlet-class>
 	    <init-param>
 	      	<param-name>applicationClassName</param-name>
-	      	<param-value>net.stellers.sakai.academicsessionadmin.tool.AcademicSessionAdminApplication</param-value>
+	      	<param-value>org.sakaiproject.acadtermmanage.tool.AcademicSessionAdminApplication</param-value>
 	    </init-param>
 		<load-on-startup>1</load-on-startup>
   	</servlet>
   	
   	<servlet-mapping>
-  		<servlet-name>sakai.term-admin</servlet-name>
+  		<servlet-name>sakai.acadtermmanage</servlet-name>
   		<url-pattern>/*</url-pattern>
   		 
 	</servlet-mapping>
@@ -40,7 +40,7 @@
 	</filter>
 	<filter-mapping>
 		<filter-name>sakai.request</filter-name>
-		<servlet-name>sakai.term-admin</servlet-name>		
+		<servlet-name>sakai.acadtermmanage</servlet-name>		
 		<dispatcher>REQUEST</dispatcher>
 		<dispatcher>FORWARD</dispatcher>
 		<dispatcher>INCLUDE</dispatcher>
